Bunch o' fixes
+ All Leaderboard and Achievement requests should work properly now
+ Switched out int64 for int32 in most situations for compatibility with UE4.21 and prior
+ Switched to using FStrings instead of int64 for UserIds, for the same reason
+ Added a couple of helper functions for conversions, in case I need to change something later on
+ Reversed the slashes on the OVR_Platform.h include so it'll compile for UE4.21 and prior

+++ OculusPlatformBP/Source/OculusPlatformBP/Private/OBPRichPresence.cpp
@@ -269,7 +269,7 @@ void UOBP_RichPresence::SetDeeplinkMessageOverride(const FString RichPresenceDee
 #endif
 }
 
-void UOBP_RichPresence::SetEndTime(const int64 RichPresenceEndTime)
+void UOBP_RichPresence::SetEndTime(const int32 RichPresenceEndTime)
 {
 #if PLATFORM_MINOR_VERSION >= 39
 	ovr_RichPresenceOptions_SetEndTime(OvrRichPresenceOptions, RichPresenceEndTime);
@@ -331,7 +331,7 @@ void UOBP_RichPresence::SetMaxCapacity(const int32 RichPresenceMaxCapacity)
 #endif
 }
 
-void UOBP_RichPresence::SetStartTime(const int64 RichPresenceStartTime)
+void UOBP_RichPresence::SetStartTime(const int32 RichPresenceStartTime)
 {
 #if PLATFORM_MINOR_VERSION >= 40
 	ovr_RichPresenceOptions_SetStartTime(OvrRichPresenceOptions, RichPresenceStartTime);
